Fixes #15731 - introduced discovery_delete_facts setting

+++ app/services/foreman_discovery/host_converter.rb
@@ -15,10 +15,12 @@ class ForemanDiscovery::HostConverter
     if set_build
       # set legacy_api flag for post_queue actions
       host.legacy_api = self.legacy_host(host)
-      # do not delete all facts (keep discovery ones)
-      host.define_singleton_method(:clear_facts) do
-        keep_ids = FactValue.where("host_id = #{host.id}").joins(:fact_name).where("fact_names.name like 'discovery_%'").pluck("fact_values.id")
-        FactValue.where("host_id = #{host.id} and id not in (?)", keep_ids).delete_all
+      # do not delete all facts (keep all facts or discovery ones when configured)
+      unless Setting['discovery_clean_facts']
+        host.define_singleton_method(:clear_facts) do
+            keep_ids = FactValue.where("host_id = #{host.id}").joins(:fact_name).where("fact_names.name like 'discovery_%'").pluck("fact_values.id")
+            FactValue.where("host_id = #{host.id} and id not in (?)", keep_ids).delete_all
+        end
       end
       # set build flag (also deletes facts)
       host.build = set_build
